refactor folderstructure
