+++ client/src/components/content/BpmnModeler/BpmnModeler.jsx
@@ -1,8 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import CamundaBpmnModeler from 'bpmn-js/lib/Modeler';
 import { Layout } from 'antd';
 import propertiesProviderModule from 'bpmn-js-properties-panel/lib/provider/camunda';
 import camundaModdleDescriptor from 'camunda-bpmn-moddle/resources/camunda';
+import PropTypes from 'prop-types'
 import { emptyBpmn } from '../../../resources/modeler/empty.bpmn';
 // eslint-disable-next-line camelcase
 import arkRPA_ModdleDescriptor from '../../../resources/modeler/modelerPropertiesExtensionRPA/ark-rpa.json';
@@ -55,6 +56,10 @@ const BpmnModeler = (props) => {
       <div className={styles['bpmn-modeler-container']} id='bpmnview' />
     </Content>
   );
+};
+
+BpmnModeler.propTypes = {
+  onModelerUpdate: PropTypes.func.isRequired,
 };
 
 export default BpmnModeler;
