+++ client/src/components/content/ModelerSidebar/ModelerSidebar.module.css
@@ -1,3 +1,14 @@
+.label-on-dark-background {
+  color: var(--colorPrimaryInvertedText);
+}
+.title {
+  margin-bottom: 0rem !important;
+  color: var(--colorPrimaryInvertedText) !important;
+}
+.properties-panel-dropdown {
+  width: 100%;
+}
+
 .sider {
   background: var(--colorPrimaryInverted2);
   padding: 1rem;
