+++ client/src/components/content/ModelerSidebar/PropertiesPanel/PropertiesPanelSections/PPOutputValueSection.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Space, Typography, Tooltip, Input } from 'antd';
 import { InfoCircleOutlined } from '@ant-design/icons';
 import corporateDesign from '../../../../../layout/corporateDesign';
-import styles from '../../PropertiesPanel.module.css';
+import styles from '../../ModelerSidebar.module.css';
 
 const { Text } = Typography;
