+++ client/src/components/content/ModelerSidebar/PropertiesPanel/PropertiesPanelSections/PPTitleSection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Typography } from 'antd';
-import styles from '../../PropertiesPanel.module.css';
+import styles from '../../ModelerSidebar.module.css';
 
 const { Text } = Typography;
