+++ client/src/components/pages/Modeler/Modeler.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Layout } from 'antd';
 import BpmnModeler from '../../content/BpmnModeler/BpmnModeler';
 import HeaderNavbar from '../../content/HeaderNavbar/HeaderNavbar';
-import ModelerSidebar from '../../content/ModelerSidebar/ModelerSidebar';
+import PropertiesPanel from '../../content/ModelerSidebar/ModelerSidebar'
 
 import 'bpmn-js/dist/assets/diagram-js.css';
 import 'bpmn-font/dist/css/bpmn-embedded.css';
@@ -24,7 +24,7 @@ const Modeler = () => {
       <HeaderNavbar selectedKey={2} />
       <Layout>
         <BpmnModeler onModelerUpdate={updateModeler} />
-        <ModelerSidebar modeler={modeler} />
+        {modeler && <PropertiesPanel modeler={modeler} />}
       </Layout>
     </>
   )
